Fix mutating avatar's FX layers w/o Modular Avatar
If Modular Avatar is installed and runs before this plugin, it will clone all the animators so it'd be safe to make changes. It doesn't make sense to rely on MA just for this functionality so include this functionality ourselves (pragmatically it probably could be argued it does make sense because I'm going to guess most users of this are going to have MA installed, but oh well let's take this road for now and we can always reverse course later)

Fixes #7, leave the "how do we know it won't touch the original files" question to adding unit tests i.e. #1