Move animator cloning into write defaults pass
If this was a larger project it may make sense to only do it once in the NDMF resolving phase, but otherwise this is a case of YAGNI. Cloning in the write defaults pass is fine. We can revisit this decision if more passes end up needing this. Ideally https://github.com/bdunderscore/ndmf/issues/111 lands and we can migrate to that instead, since it's already duplicate work for this and Modular Avatar to both do a deep clone of avatar FX