Integrate Blendshape Viewer into package (#25)
This replaces the previous submodule method of pulling the package into this repository, and replaces the VPM package dependency with symlinked pre-patched files within this repository.

These patches are saved separately in a config that allows the patching process to be performed by a bot, as well as being manually reproducible.

Here we do the minimal amount of effort required to replace the submodule, as well as renaming namespaces etc. to try and avoid collisions with upstream. Further refactoring and fixes, such as changing `public` declarations to `internal`, are left to future changes.

---------

Co-authored-by: enitimeago-bot[bot] <159699746+enitimeago-bot[bot]@users.noreply.github.com>