Fix editor breaking when entering play mode (#38)
The EditorWindow is effectively reset when entering/leaving play mode, so we need to pull in data again.

Currently we just have a hack that detects if the reference to the blend shape mappings component turns into null and stop rendering the window. But this persists, so the user has to reload the window.

However the instance ID of an object can be easily persisted across editor/play mode, so use that. (Brief investigation was made into `[SerializedField]`, but this didn't work and was not pursued further.

To make things easier on ourselves don't allow editing when in play mode, instead show the GUI in a read-only mode.